README
Spatio Temporal analysis involves analysing the coverage of various cellular networks in and around Stony Brook University. 
We present the details of the analysis and the results of our experiment on the consistency of the network coverage of 
various cellular networks, across space and time by developing an android application which records the required data. 
We have collected information from various devices across the Stony Brook University premises and we have analysed 
the captured data to discover potential problems.